bugfix: fix exec not exit with correct exit code
add ContainerExecInspect client api, refactor container
part in client package.
make exec exit with correct exit code.

Fixes: #344.
