+++ client/interface.go
@@ -28,6 +28,7 @@ type ContainerAPIClient interface {
 	ContainerAttach(ctx context.Context, name string, stdin bool) (net.Conn, *bufio.Reader, error)
 	ContainerCreateExec(ctx context.Context, name string, config *types.ExecCreateConfig) (*types.ExecCreateResp, error)
 	ContainerStartExec(ctx context.Context, execid string, config *types.ExecStartConfig) (net.Conn, *bufio.Reader, error)
+	ContainerExecInspect(ctx context.Context, execid string) (*types.ContainerExecInspect, error)
 	ContainerGet(ctx context.Context, name string) (*types.ContainerJSON, error)
 	ContainerRename(ctx context.Context, id string, name string) error
 	ContainerRestart(ctx context.Context, name string, timeout string) error
